Moved shader-related functions to shaders.py
* Also, tested calling as an operator => still works fine
* Added a couple of comments
