+++ export_gms_vtx_buffer.py
@@ -5,7 +5,8 @@ from struct import (
     )
 
 # Mesh-like objects (the ones that can be converted to mesh)
-meshlike_types = {'MESH', 'CURVE', 'SURFACE', 'FONT', 'META'}
+MESHLIKE_TYPES = {'MESH', 'CURVE', 'SURFACE', 'FONT', 'META'}
+
 def triangulated_mesh_from_object(obj):
     """Important: use to_mesh_clear to free the mesh generated by this function"""
     mod_tri = obj.modifiers.new('triangulate_for_export', 'TRIANGULATE')
@@ -17,6 +18,7 @@ def triangulated_mesh_from_object(obj):
     m.calc_normals_split()
     obj.modifiers.remove(mod_tri)
     return m
+
 
 def write_object_ba(scene, obj, desc, ba, frame, reverse_loop, apply_transforms):
     """Traverse the object's mesh data at the given frame and write to the
@@ -31,16 +33,22 @@ def write_object_ba(scene, obj, desc, ba, frame, reverse_loop, apply_transforms)
                 for offset, attr_blen, fmt, index, func, args in occurrences:   # Each occurrence's data (tuple assignment!)
                     ind = byte_pos + offset
                     if id.startswith('ShaderNode'):
+                        # Write the input's default value for shader nodes
+                        # TODO Trace back to value if a node is attached (if doable)
                         val = node.inputs[prop].default_value
                     else:
+                        # Simply lookup the property on the node
                         val = getattr(node, prop)
                     
+                    # Pass through pre-process function
                     if func != None:
                         if args == "":
                             val = func(val, ctx=ctx)
                         else:
                             val = func(val, ctx=ctx, args=json.loads(args))
                     
+                    # Pack as bytes according to the chosen data type
+                    # and write to correct position in the right bytearray
                     val_bin = pack(fmt, val) if len(fmt) == 1 else pack(fmt, *val[:len(fmt)])
                     ba[frame-index][ind:ind+attr_blen] = val_bin
 
@@ -71,7 +79,7 @@ def write_object_ba(scene, obj, desc, ba, frame, reverse_loop, apply_transforms)
 
                 if mat.use_nodes:
                     # Simply loop through all of this material's nodes
-                    # (Todo: could be optimized quite a bit..)
+                    # (TODO could be optimized quite a bit..)
                     for node in mat.node_tree.nodes:
                         fetch_attribs(desc, node, ba, ba_pos, frame, ctx)
                 
@@ -211,7 +219,7 @@ def export(context, filepath,
 
     # Which models to export
     object_selection = context.selected_objects if selection_only else context.scene.objects
-    mesh_selection = [obj for obj in object_selection if obj.type in meshlike_types]    # TODO Does this break morphs?
+    mesh_selection = [obj for obj in object_selection if obj.type in MESHLIKE_TYPES]    # TODO Does this break morphs?
     for i, obj in enumerate(mesh_selection): obj.batch_index = i   # Guarantee a predictable batch index
 
     # Support alternative extension for model files
@@ -320,15 +328,6 @@ def export(context, filepath,
         node_based_materials = [mat for mat in materials if mat.use_nodes]
         for mat in node_based_materials:
             ntree = mat.node_tree
-
-            """
-            # More advanced traversal of tree
-            output_node = [n for n in mat.node_tree.nodes if len(n.outputs) == 0]
-
-            node = output_node
-            while node:
-                node = node.inputs[0].links[0].from_node    # Phew...
-            """
 
             if len(ntree.nodes) > 1:    # Quite a couple of happy assumptions we make here...
                 tex_node = [n for n in ntree.nodes if n.type == 'TEX_IMAGE']
